#81 💵 fix : 회원가입 유효성 검사 오류 수정
